1. Reinstated progress bar 2. More renames
git-svn-id: svn+ssh://dev.open-bio.org/home/svn-repositories/biojava/RCSB Viewers/branches/0820@5834 7c6358e6-4a41-0410-a743-a5b2a554c398

+++ RCSB MBT Libs/Structure Model/org/rcsb/mbt/model/util/StatusEvent.java
@@ -106,6 +106,6 @@ public class StatusEvent
 	/**
 	 * The progress percentage (0.0-1.0).
 	 */
-	public float percent = 0.0f;
+	public int percent = 0;
 }
